rename package
-renamed package to be able to get releasable apk for google playstore
