+++ app/google-services.json
@@ -10,7 +10,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:196086420703:android:8276ff67bbddad5580c56a",
         "android_client_info": {
-          "package_name": "com.example.finalProject"
+          "package_name": "com.taskApp.finalProject"
         }
       },
       "oauth_client": [
@@ -18,7 +18,7 @@
           "client_id": "196086420703-86ar9nniglg5kv8crl15vni6cfa64732.apps.googleusercontent.com",
           "client_type": 1,
           "android_info": {
-            "package_name": "com.example.finalProject",
+            "package_name": "com.taskApp.finalProject",
             "certificate_hash": "3ad298d90fa59fc47938af67008210e4597e014f"
           }
         },
