+++ app/src/main/res/layout/content_main.xml
@@ -16,7 +16,7 @@
         android:layout_height="wrap_content"
         android:layout_marginBottom="148dp"
         android:onClick="launchCompletedTaskActivity"
-        android:text="View Completed Tasks"
+        android:text="@string/view_completed_tasks"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="@+id/guideline2"
@@ -39,14 +39,15 @@
         app:layout_constraintTop_toBottomOf="@+id/description_tv"
         app:layout_constraintVertical_bias="0.686"
         app:srcCompat="@drawable/add_image"
-        tools:srcCompat="@drawable/add_image" />
+        tools:srcCompat="@drawable/add_image"
+        android:contentDescription="@string/todo" />
 
     <Button
         android:id="@+id/completeTaskBtn"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginBottom="148dp"
-        android:text="Complete Task"
+        android:text="@string/complete_task"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toStartOf="@+id/guideline2"
         app:layout_constraintStart_toStartOf="parent"
@@ -59,10 +60,10 @@
         android:layout_height="114dp"
         android:layout_marginTop="32dp"
         android:cursorVisible="false"
-        android:drawableRight="@drawable/ic_dialog_close_light"
+        android:drawableEnd="@drawable/ic_dialog_close_light"
         android:ems="10"
         android:gravity="start|top"
-        android:hint="Enter a description about your completed tasks here (optional)"
+        android:hint="@string/enter_a_description_about_your_completed_tasks_here_optional"
         android:inputType="textMultiLine"
         android:textAlignment="center"
         app:endIconMode="clear_text"
@@ -76,7 +77,7 @@
         android:id="@+id/time_tv"
         android:layout_width="351dp"
         android:layout_height="33dp"
-        android:text="Elapsed Time:"
+        android:text="@string/elapsed_time"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.533"
         app:layout_constraintStart_toStartOf="parent"
@@ -90,7 +91,7 @@
         android:layout_marginTop="8dp"
         android:layout_marginEnd="8dp"
         android:onClick="launchWebViewActivity"
-        android:text="Tap to contribute new task ideas"
+        android:text="@string/tap_to_contribute_new_task_ideas"
         android:textColor="#3F51B5"
         android:textSize="12sp"
         app:layout_constraintBottom_toTopOf="@+id/task_tv"
@@ -118,7 +119,7 @@
         android:layout_marginStart="32dp"
         android:layout_marginTop="8dp"
         android:onClick="addTask"
-        android:text="Get task:"
+        android:text="@string/get_task"
         app:layout_constraintBottom_toTopOf="@+id/task_tv"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
@@ -129,7 +130,7 @@
         android:layout_width="352dp"
         android:layout_height="78dp"
         android:layout_marginTop="40dp"
-        android:text="Press button to get a task"
+        android:text="@string/press_button_to_get_a_task"
         android:textSize="24sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.508"
